Revert "Merge das classes"
This reverts commit f58476b6da7050d6797aed4079c584768dbe2eb7.

+++ src/main/java/br/com/newObjects/model/entities/classes/NovoPedido.java
@@ -12,7 +12,7 @@ public class NovoPedido {
     public static Pedido novoPedido() {
 
         Pedido pedido = new Pedido(NovoCliente.novoCliente(),
-                FormaPagamento.novaFormaPagamento(new Random().nextInt(5) + 1), new Random().nextInt(10));
+                FormaPagamento.novaFormaPagamento(new Random().nextInt(5) + 1), new Random().nextFloat(10));
 
         pedido.setItensPedido(
                 NovoItemPedido.novaListaItensPedido(pedido, NovoProdutoFornecedor.novaListaProdutosFornecedor()));
@@ -25,6 +25,6 @@ public class NovoPedido {
 
     public static Pedido novoPedidoSimples() {
         return new Pedido(NovoCliente.novoCliente(), FormaPagamento.novaFormaPagamento(new Random().nextInt(5) + 1),
-                new Random().nextInt(100) + 1);
+                new Random().nextFloat(100) + 1);
     }
 }
